Django_Practice5/requirements.txt
asgiref==3.8.1
Django==5.2
django-allauth==65.7.0
django-filter==25.1
sqlparse==0.5.3
